crew/mcp_server/requirements.txt
crewai>=0.11.0
langchain-openai>=0.0.2
openai>=1.3.5
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
starlette>=0.27.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
supabase>=1.0.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.4.0